refactor: revert collections structure and mark Graph as experimental
- Move Graph.ts back to src/ directory
- Mark Graph module as @experimental in both index.ts and Graph.ts
- Update changeset to reflect experimental status and correct example
- Verify changeset example works correctly
