+++ packages/effect/src/Graph.ts
@@ -1,17 +1,18 @@
 /**
+ * @experimental
  * @since 3.0.0
  */
 
-import * as Data from "../Data.js"
-import * as Equal from "../Equal.js"
-import { dual } from "../Function.js"
-import * as Hash from "../Hash.js"
-import type { Inspectable } from "../Inspectable.js"
-import { format, NodeInspectSymbol } from "../Inspectable.js"
-import * as Option from "../Option.js"
-import type { Pipeable } from "../Pipeable.js"
-import { pipeArguments } from "../Pipeable.js"
-import type { Mutable } from "../Types.js"
+import * as Data from "./Data.js"
+import * as Equal from "./Equal.js"
+import { dual } from "./Function.js"
+import * as Hash from "./Hash.js"
+import type { Inspectable } from "./Inspectable.js"
+import { format, NodeInspectSymbol } from "./Inspectable.js"
+import * as Option from "./Option.js"
+import type { Pipeable } from "./Pipeable.js"
+import { pipeArguments } from "./Pipeable.js"
+import type { Mutable } from "./Types.js"
 
 /**
  * Safely get a value from a Map, returning an Option.
