+++ packages/effect/src/index.ts
@@ -351,9 +351,10 @@ export * as Function from "./Function.js"
 export * as GlobalValue from "./GlobalValue.js"
 
 /**
+ * @experimental
  * @since 3.0.0
  */
-export * as Graph from "./collections/Graph.js"
+export * as Graph from "./Graph.js"
 
 /**
  * @since 2.0.0
